feat(stocks): display relative date on stock cards — closes #62
